requirements.txt
# WebClaw DeepAgent - 核心依赖（精简版）

# === 核心框架 ===
langchain>=0.3.0,<1.0
langchain-community>=0.3.0,<1.0
langchain-openai>=0.2.0,<1.0
langgraph>=0.2.0,<2.0

# === 浏览器自动化 ===
playwright>=1.40.0,<2.0

# === 数据模型 ===
pydantic>=2.0.0,<3.0

# === 日志 ===
loguru>=0.7.0,<1.0

# === 环境变量 ===
python-dotenv>=1.0.0,<2.0

# === 测试 ===
pytest>=7.0.0,<9.0
pytest-asyncio>=0.21.0,<1.0

# === 可选：向量数据库（阶段5+）===
# chromadb>=0.4.0,<1.0

# === 可选：RL训练（阶段5+）===
# gymnasium>=0.29.0,<1.0
# torch>=2.0.0,<3.0